==> 01_Data Import.ipynb <==
# THE DATA


IMDB Provides Several Files with varied information for Movies, TV Shows, Made for TV Movies, etc.

Overview/Data Dictionary: https://www.imdb.com/interfaces/

Downloads page: https://datasets.imdbws.com/

I will use files below for this project: 

* title.basics.tsv.gz
* title.ratings.tsv.gz
* title.akas.tsv.gz

#import libraries
import pandas as pd
import numpy as np

#download Datasets
basics = pd.read_csv('https://datasets.imdbws.com/title.basics.tsv.gz', 
                     sep='\t', low_memory=False)
basics.head(2)

basics.info()

ratings = pd.read_csv('https://datasets.imdbws.com/title.ratings.tsv.gz', 
                      sep='\t', low_memory=False)
ratings.head(2)

ratings.info()

akas = pd.read_csv('https://datasets.imdbws.com/title.akas.tsv.gz', 
                   sep='\t', low_memory=False)
akas.head(2)

akas.info()

#change null values encoding from \N to np.nan
basics = basics.replace({'\\N':np.nan}) 
ratings = ratings.replace({'\\N':np.nan}) 
akas = akas.replace({'\\N':np.nan}) 

#eliminate movies that are null for runtimeMinutes and genre
basics = basics.dropna(subset = ['runtimeMinutes', 'genres','startYear'])

#keep only titleType = Movie
basics = basics.loc[basics['titleType'] == 'movie']

#keep startYear 2000-2022
basics['startYear'] = basics['startYear'].astype(int)
basics = basics.loc[(basics['startYear'] >= 2000) 
                    & (basics['startYear'] <= 2022)]

basics.head()

#exclude movies that are included in the documentary category
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]

#filter to have only US movies
akas = akas.loc[akas['region'] == 'US']

#filter the basics table down to only include the US by using the filter akas dataframe
keepers = basics['tconst'].isin(akas['titleId'])

basics = basics[keepers]
basics.head()

#create new folder for Data
import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

#save dataframes to Compressed .csv.gz files
basics.to_csv('Data/basics.csv.gz',compression='gzip',index=False)
ratings.to_csv('Data/ratings.csv.gz',compression='gzip',index=False)
akas.to_csv('Data/akas.csv.gz',compression='gzip',index=False)

#open saved file and preview again
basics = pd.read_csv("Data/basics.csv.gz", low_memory = False)
basics.head()
