==> 02_Adding financial data.ipynb <==
**Extract more Data from TMDB!**

* There is no financial information included in the IMDB data (e.g. budget or revenue).

* This will be a major roadblock when attempting to analyze which movies are successful and must be addressed before you will be able to determine which movies are successful.

* The Movie Database (TMDB) as a great source of financial data (https://www.themoviedb.org/) and offers a free API for programmatic access to their data!

* My stakeholder wants you to extract the budget, revenue, and MPAA Rating (G/PG/PG-13/R), which is also called "Certification".

**Specifications**

* My stakeholder would like me to extract and save the results for movies that meet all of the criteria established in the first part of the project.

* As a proof-of-concept, they requested I perform a test extraction of movies that started in 2000 or 2001

* Each year should be saved as a separate .csv.gz file

#import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm_notebook

import json
with open('/Users/oleksandraaliyeva/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()

import tmdbsimple as tmdb
tmdb.API_KEY =  login['api-key']

#create folder for data
import os
FOLDER = "Data"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)

#creat instance for years
YEARS_TO_GET = list(range(2000,2022))

#function to get movie rating
def get_movie_with_rating(movie_id): 
    
    #get the movie object for the current id
    movie = tmdb.Movies(movie_id)
    
    #save the .info .releases dictionaries
    movie_info = movie.info()
    releases = movie.releases()
    
    #loop through countries in releases
    for c in releases['countries']:
        #if the country abbreviation==US
        if c['iso_3166_1'] =='US':
            #save a "certification" key in the info dict with the certification
            movie_info['certification'] = c['certification']
            
    return movie_info

#function to create a new json file 
def write_json(new_data, filename): 
        """Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
        
        with open(filename,'r+') as file:
            #first we load existing data into a dict.
            file_data = json.load(file)
            
            #choose extend or append
            if (type(new_data) == list) & (type(file_data) == list):
                file_data.extend(new_data)
            else:
                file_data.append(new_data)
                
        #sets file's current position at offset.
            file.seek(0)
            
        #convert back to json.
            json.dump(file_data, file)

#load in the dataframe from project part 1 as basics:
basics = pd.read_csv('Data/basics.csv.gz')

#start of OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET,desc='YEARS',position=0):
    
    #defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    
    #check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    
    #if it does not exist: create it
    if file_exists == False:
        
        #save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
    
    #saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    
    #saving movie ids to list
    movie_ids = df['tconst'].copy()
   
    #load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
    
    #INNER Loop
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        #attempt to retrieve then data for the movie id
        try:
            temp = get_movie_with_rating(movie_id) 
        
            #append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            
            #short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        #if it fails,  make a dict with just the id and None for certification.
        except Exception as e:
            continue
      
    #open json file and save as csv
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

#read csv file for 2000
y_2000 = pd.read_csv('Data/final_tmdb_data_2000.csv.gz')

#drop first row and check the head
y_2000 = y_2000.iloc[1: , :]
y_2000.head()

#read csv file for 2000
y_2001 = pd.read_csv('Data/final_tmdb_data_2001.csv.gz')

#drop first row and check the head
y_2001 = y_2001.iloc[1: , :]
y_2001.head()

#concatenate the previous results and new results. 
merged = pd.concat([y_2000, y_2001],
                      ignore_index=True)
display(merged.head(3), merged.tail(3))

merged.to_csv("Data/tmdb_results_combined_2000_2001.csv.gz", 
              compression="gzip", index=False)

#add the data from years 2002-2022
YEARS_TO_GET = list(range(2002, 2022))
for YEAR in YEARS_TO_GET:
    df = pd.read_csv(f"Data/final_tmdb_data_{YEAR}.csv.gz", 
                     lineterminator='\n')
    merged = pd.concat([merged, df])

#create one file with all data from API
merged.to_csv("Data/tmdb_results_combined.csv.gz", compression="gzip", 
              index=False)
