==> 04_Create a MySQL database .ipynb <==
**Specifications - Database**

My stakeholder want me to take the data I have been cleaning and collecting previously in the project, and create a MySQL database for them.

#import libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
connection_str = "mysql+pymysql://root:952175aK@localhost/movies"

#create engine
engine = create_engine(connection_str)

#check if database exists
if database_exists(connection_str) == False: create_database(connection_str)
else: print('The database already exists.')

**Required Transformation steps:**

Normalize Genre:

* Convert the single string of genres from title basics into 2 new tables.

1. title_genres: with the columns:
    * tconst
    * genre_id

2. genres:

    * genre_id
    * genre_name

#load data 
basics = pd.read_csv("Data/basics.csv.gz", low_memory = False)
basics.head()

#creat a new column with a string transformed to a list
basics['genres_split'] = basics['genres'].str.split(',')
basics.head()

#exploding the column of lists
exploded_geners = basics.explode('genres_split')
exploded_geners[['tconst','genres','genres_split']].head(5)

#saving the unique values from the exploded column and sorting
unique_geners = sorted(exploded_geners['genres_split'].dropna().unique())
unique_geners

#creat new dataframe
title_genres = exploded_geners[['tconst', 'genres_split']].copy()
title_genres.head()

#make the genre map dictionary
genre_ints = range(len(unique_geners))
genres_map = dict(zip(unique_geners, genre_ints))
genres_map

#make new culumn genre_id
title_genres['genre_id'] = title_genres['genres_split'].map(genres_map)

#drop column 'genres_split'
title_genres = title_genres.drop(columns='genres_split')
title_genres.head()

#create new dataframe for genres
genre_lookup = pd.DataFrame({'genre_name': genres_map.keys(), 
                            'genre_id': genres_map.values()})
genre_lookup.head()

**Discard unnecessary information:**

For the title basics table, drop the following columns:
* "original_title" (we will use the primary title column instead)
* "isAdult" ("Adult" will show up in the genres so this is redundant information).
* "titleType" (every row will be a movie).
* "genres" and other variants of genre (genre is now represented in the 2 new tables described above.

#drop unnecessary columns
basics = basics.drop(columns = ['originalTitle','isAdult', 'titleType', 
                                'genres', 'genres_split'])
basics

**Export basics Table to MySQL**

#check type of Data
basics.dtypes

from sqlalchemy.types import *

#сalculate max string lengths for object columns
key_len = basics['tconst'].fillna('').map(len).max()
title_len = basics['primaryTitle'].fillna('').map(len).max()

#create a schema dictonary using Sqlalchemy datatype objects
basics_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Integer(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

#create table title_basics
basics.to_sql('title_basics',engine,dtype=basics_schema,if_exists='replace',
          index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

#check sql table
q = """
SELECT *
FROM title_basics
LIMIT 5
"""
pd.read_sql(q, engine)

**Export title_rating Table to MySQL**

#load data of ratings
ratings = pd.read_csv('Data/ratings.csv.gz', low_memory = False)

#check type of Data
ratings.dtypes

#сalculate max string lengths for object columns
key_len = ratings['tconst'].fillna('').map(len).max()

#create a schema dictonary using Sqlalchemy datatype objects
ratings_schema = {
    "tconst": String(key_len+1),
    "averageRating": Float(),
    'numVotes':Integer()}

#create table title_ratings
ratings.to_sql('title_ratings',engine,dtype=ratings_schema,if_exists='replace',
               index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

#check sql table
q = """
SELECT *
FROM title_ratings
LIMIT 5
"""
pd.read_sql(q, engine)

**Export genres Table to MySQL**

#check type of Data
genre_lookup.dtypes

#сalculate max string lengths for object columns
genre_name_len = genre_lookup['genre_name'].fillna('').map(len).max()

#create a schema dictonary using Sqlalchemy datatype objects
genre_lookup_schema = {
    "genre_name": Text(genre_name_len+1),
    'genre_id':Integer()}

#create table genres
genre_lookup.to_sql('genres',engine,dtype=genre_lookup_schema,
                    if_exists='replace',index=False)

engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

#check sql table
q = """
SELECT *
FROM genres
LIMIT 5
"""
pd.read_sql(q, engine)

**Export title_genres Table to MySQL**

#check type of Data
title_genres.dtypes

#сalculate max string lengths for object columns
key_len = title_genres['tconst'].fillna('').map(len).max()

#create a schema dictonary using Sqlalchemy datatype objects
title_genre_schema = {
    "tconst": String(key_len+1), 
    'genre_id':Integer()}

#save to sql with dtype and index=False
title_genres.to_sql('title_genres',engine,dtype=title_genre_schema,
                    if_exists='replace',index=False)

#check sql table
q = """
SELECT *
FROM title_genres
LIMIT 5
"""
pd.read_sql(q, engine)

**Export tmdb_data Table to MySQL**

#load financial data
tmdb_data = pd.read_csv('Data/tmdb_results_combined.csv.gz', low_memory = False, lineterminator='\n')

#leave columns needed for analisys
tmdb_data = tmdb_data[['imdb_id', 'revenue', 'budget', 'certification']]
tmdb_data.head()

tmdb_data.info()

#drop all rows with missing information
tmdb_data = tmdb_data.loc[tmdb_data['imdb_id'] != '0']

#check for duplicates
tmdb_data.duplicated().sum()

#drop duplicates
tmdb_data.drop_duplicates(inplace=True)

#check for duplicates
tmdb_data.duplicated().sum()

#drop movies with 0 budget.
tmdb_data = tmdb_data.loc[tmdb_data['revenue'] > 0]

#check column certification for unique values
tmdb_data['certification'].value_counts()

# fix extra space certs
tmdb_data['certification'] = tmdb_data['certification'].str.strip()
tmdb_data['certification'].value_counts(dropna=False)

#check type of Data
tmdb_data.dtypes

#сalculate max string lengths for object columns
key_len = tmdb_data['imdb_id'].fillna('').map(len).max()
certification_len = tmdb_data['certification'].fillna('').map(len).max()

#create a schema dictonary using Sqlalchemy datatype objects
tmdb_schema = {
    "imdb_id": String(key_len+1),
    "revenue": Float(),
    'budget':Float(),
    'certification': Text(certification_len+1)}

#create table tmdb_data
tmdb_data.to_sql('tmdb_data',engine,dtype=tmdb_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

#check sql table
q = """
SELECT *
FROM tmdb_data
LIMIT 5
"""
pd.read_sql(q, engine)

**Checking DATABASE**

#check the database
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
